# openapi.yaml
openapi: 3.0.3
info:
  title: Wedding RSVP API
  version: 1.0.0
paths:
  /rsvp:
    post:
      summary: Create an RSVP record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullname, email, status, guests, place]
              properties:
                fullname: { type: string }
                email: { type: string, format: email }
                status: { type: string, enum: [yes, no] }
                guests: { type: integer, minimum: 1, maximum: 10 }
                place: { type: string, example: "Tiệc Nhà Trai" }
                message: { type: string }
      responses:
        "201": { description: Created }
        "400": { description: Bad Request }
        "500": { description: Server Error }
  /addWish:
    post:
      summary: Add a wish to loved ones
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullname, relationship, wish_msg, predefined_wish]
              properties:
                fullname: { type: string }
                relationship: { type: string }
                wish_msg: { type: string }
                predefined_wish: { type: string }
                createdAt: { type: string }
      responses:
        "201": { description: Created }
        "400": { description: Bad Request }
        "500": { description: Server Error }
  /getWishes:
    get:
      summary: Get a list of wishes
      description: Retrieve all wishes stored in the Airtable "Wishes" table.
      operationId: getWishes
      parameters:
        - name: Authorization
          in: header
          required: true
          description: Airtable API key in the format `Bearer YOUR_API_KEY`
          schema:
            type: string
      responses:
        "200":
          description: A list of wishes
          content:
            application/json:
              schema:
                type: object
                properties:
                  records:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: recXXXXXXXXXXXX
                        fields:
                          type: object
                          properties:
                            FullName:
                              type: string
                              example: John Doe
                            WishMessage:
                              type: string
                              example: I wish for peace and happiness.
                            PredefinedWish:
                              type: string
                              example: Happy Birthday!
                            CreatedAt:
                              type: string
                              example: Happy Birthday!
                        createdTime:
                          type: string
                          format: date-time
                          example: 2025-10-28T12:34:56.000Z
        "401":
          description: Unauthorized – missing or invalid API key.
        "500":
          description: Internal server error.